Variables can be used as parameters in the SHapes and methods
In this commit, The variables declared in the application can be used as parameters.

+++ Write_code_to_draw/Circle.cs
@@ -17,7 +17,7 @@ namespace Write_code_to_draw
     public class Circle : Shapes
     {
 
-        int radius;
+        float radius;
         Pen pen;
         SolidBrush brush;
 
@@ -27,7 +27,7 @@ namespace Write_code_to_draw
         /// </summary>
         /// <param name="param"></param> the integer array of the parameters given by the user to be constructed in the application
         /// <param name="Default_Pen"></param> Pen for drawing the shape if the fill is off
-        public void GetValue(int[] param,Pen Default_Pen)
+        public void GetValue(float[] param,Pen Default_Pen)
         {
             this.radius = param[0];
             this.pen = Default_Pen;
@@ -38,7 +38,7 @@ namespace Write_code_to_draw
         /// </summary>
         /// <param name="param"></param> the integer array of the parameters given by the user to be constructed in the application
         /// <param name="Default_Pen"></param> Pen for drawing the shape if the fill is off
-        public void GetValue(int[] param, SolidBrush Default_Brush)
+        public void GetValue(float[] param, SolidBrush Default_Brush)
         {
             this.radius = param[0];
             this.brush= Default_Brush;
