+++ Write_code_to_draw/Rectangle.cs
@@ -13,7 +13,7 @@ namespace Write_code_to_draw
     public class Rectangle :Shapes
     {
 
-        int side1, side2;
+        float side1, side2;
         Pen pen;
         SolidBrush brush;
         /// <summary>
@@ -21,7 +21,7 @@ namespace Write_code_to_draw
         /// </summary>
         /// <param name="param"></param> the integer array of the parameters given by the user to be constructed in the application
         /// <param name="Default_Pen"></param> Pen for drawing the shape if the fill is off
-        public void GetValue(int[] param, Pen Default_Pen)
+        public void GetValue(float[] param, Pen Default_Pen)
         {
             this.side1 = param[0];
             this.side2 = param[1];
@@ -33,7 +33,7 @@ namespace Write_code_to_draw
         /// </summary>
         /// <param name="param"></param> the integer array of the parameters given by the user to be constructed in the application
         /// <param name="Default_Brush"></param> Brush to fill inside the shape if the fill is on
-        public void GetValue(int[] param, SolidBrush Default_Brush)
+        public void GetValue(float[] param, SolidBrush Default_Brush)
         {
 
             this.side1 = param[0];
