+++ Write_code_to_draw/Shapes.cs
@@ -28,14 +28,14 @@ namespace Write_code_to_draw
         /// </summary>
         /// <param name="param"></param> the integer array of the parameters given by the user to be constructed in the application
         /// <param name="Default_Pen"></param> Pen for drawing the shape if the fill is off
-        void GetValue(int[] parameters,Pen Default_Pen);
+        void GetValue(float[] parameters,Pen Default_Pen);
 
         /// <summary>
         /// Getts the value and sets the parameters of any shape passed into it
         /// </summary>
         /// <param name="param"></param> the integer array of the parameters given by the user to be constructed in the application
         /// <param name="Default_Pen"></param> Pen for drawing the shape if the fill is off
-        void GetValue(int[] parameters, SolidBrush Default_Brush);
+        void GetValue(float[] parameters, SolidBrush Default_Brush);
 
     }
 }
